debundle: split selector_codemod.rs by form
Break the ~4.2k-line selector_codemod.rs into per-form modules so per-form
work can fan out without merge contention on one giant file (readoff
minimization plan backlog item 4, split half):

- render.rs: the form-agnostic AST-holing engine (`hole_expr` / `hole_stmt` /
  `hole_object*` / `hole_function` / ...), anchor-span helpers, node builders,
  `emit_selector`, and `holes_present`.
- regex_anchor.rs: `STR_LITERAL_MATCHING_RE` pattern derivation, the
  `RegexAnchorSubstitution` post-pass, and the `accepted_regex_anchors` gate.
- minimize/: per-form minimizers — function.rs, class.rs, object.rs, var.rs,
  group.rs — over shared helpers in minimize/mod.rs (`render_via_read_off`,
  `finish_minimized_selector`, `render_var_slots`, `hole_var_init_padded`).

selector_codemod.rs keeps the CLI/YAML orchestration, the chunk index, the
anti-unification grouping, `prove_synthesized_selector`, the matcher-query
helpers, and the exact (non-holed) source renderers. Pure code move: the public
API (`run_selector_codemod`, `SelectorCodemodConfig`/`Rewrite`,
`render_selector_codemod_text`) is unchanged; items used across the new module
boundaries are now `pub(crate)`, everything else stays private.

BAZEL_TEST_INVOCATIONS=buildbuddy:01ad1701-5f7b-4b64-bd1c-5cb511d649f1
