+++ devinfra/js/debundle/selector_codemod.rs
@@ -15,7 +15,6 @@ use std::fs;
 use std::path::{Path, PathBuf};
 
 use anyhow::{Context, Result, bail};
-use readoff_render::kept_spans_for_anchor_set;
 use selector_candidate_index::SelectorCandidateIndex;
 use serde::Serialize;
 use serde_yaml::Value;
@@ -23,17 +22,28 @@ use shape_index::ShapeIndex;
 use source_match::BodyIndexFilter;
 use spec::{SourceMatch, SourceMatchIdentifierMode};
 use spec_modules::{collect_module_files, is_module_yaml, module_path_from_file};
-use swc_common::{BytePos, DUMMY_SP, Span, Spanned, SyntaxContext};
+use swc_common::{BytePos, DUMMY_SP, Span, Spanned};
 use swc_ecma_ast::*;
 use swc_ecma_visit::{Visit, VisitMut, VisitMutWith, VisitWith};
 
 // Hole keyword spellings come from `source_match_holes` so the minimizer
 // emits exactly the tokens the matcher resolves.
 use source_match_holes::{
-    ANYTHING_HOLE_KEYWORD, ARGS_HOLE_KEYWORD, CASE_REST_HOLE_KEYWORD, CLASS_REST_HOLE_KEYWORD,
-    DECLARATORS_HOLE_KEYWORD, EXPR_HOLE_KEYWORD, OBJECT_PROPS_HOLE_KEYWORD, STMT_HOLE_KEYWORD,
-    STMT_LIST_HOLE_KEYWORD, STRING_LITERAL_REGEX_PREDICATE,
+    ANYTHING_HOLE_KEYWORD, ARGS_HOLE_KEYWORD, CLASS_REST_HOLE_KEYWORD, DECLARATORS_HOLE_KEYWORD,
+    EXPR_HOLE_KEYWORD, OBJECT_PROPS_HOLE_KEYWORD, STMT_HOLE_KEYWORD,
 };
+
+// The selector minimizer is split by form: the AST-holing engine (`render`),
+// the regex-over-string-literal anchors (`regex_anchor`), and the per-form
+// minimizers (`minimize::{function,class,object,var,group}`).
+mod minimize;
+mod regex_anchor;
+mod render;
+
+use crate::minimize::{
+    minimize_class_selector, minimize_function_selector, minimize_var_group_selector,
+};
+use crate::render::holes_present;
 
 #[derive(Debug, Clone, Copy, Serialize, PartialEq, Eq)]
 #[serde(rename_all = "snake_case")]
@@ -1496,664 +1506,6 @@ fn synthesize_specialized_var_selector(
     minimize_var_group_selector(index, var, decl, targets)
 }
 
-// ===========================================================================
-// Selector minimizer (read-off based).
-//
-// A selector is the target rendered with a *retention set*: the byte spans of
-// the concrete tokens (literals, member/property names, callees, object keys)
-// the selector pins. A node renders concretely iff a kept span lies inside it;
-// every other position is holed — `ANYTHING` for a bare expression, and the
-// run holes `STMT_LIST` / `OBJECT_PROPS` / `CLASS_REST` for dropped statement /
-// object-property / class-member runs.
-//
-// Single targets (function, class, object, and non-object var) read their
-// minimal anchor set off the chunk-wide shape index (`render_via_read_off` /
-// `try_object_read_off` / `try_var_read_off`): the index ranks each candidate
-// feature by selective × stable, so the chosen anchors are sparse and
-// rebuild-robust, and the production matcher proves the rendered selector
-// resolves uniquely (gate 1). A target the read-off cannot single out returns
-// `None` and is reported as debt — never a full-AST pin.
-//
-// Multi-target var binding groups read off per slot (`try_var_group_read_off`):
-// each target declarator slot reads its minimal anchor off the shape index
-// (restricted to the slot) plus a slot-aware greedy (`slot_minimal_anchors`),
-// the per-slot kept spans union, and the binding-group matcher proves the tuple.
-// A keep-shallow path (`minimize_var_group_selector`) remains as the fallback for
-// groups whose per-slot single-binding view cannot single a slot out (a value
-// shared across sibling statements that resolves only as a tuple): it keeps each
-// slot's direct shallow literals, escalating to structural/deeper anchors only
-// until the group resolves, proven through the binding-group matcher as a
-// resolves-uniquely oracle; it may over-pin rather than run an exact-minimum
-// cover (near-minimal is the accepted target).
-// ===========================================================================
-
-/// `(lo, hi)` byte offsets of a retained concrete token.
-type AnchorSpan = (u32, u32);
-
-const MAX_MINIMIZER_ANCHORS: usize = 64;
-
-fn span_key(span: Span) -> AnchorSpan {
-    (span.lo.0, span.hi.0)
-}
-
-fn node_holds_anchor(node: Span, anchor: AnchorSpan) -> bool {
-    node.lo.0 <= anchor.0 && anchor.1 <= node.hi.0
-}
-
-fn node_retains_any(node: Span, kept: &BTreeSet<AnchorSpan>) -> bool {
-    kept.iter().any(|anchor| node_holds_anchor(node, *anchor))
-}
-
-fn ident_node(name: &str) -> Ident {
-    Ident::new_no_ctxt(name.into(), DUMMY_SP)
-}
-
-fn anything_expr() -> Expr {
-    Expr::Ident(ident_node(ANYTHING_HOLE_KEYWORD))
-}
-
-fn stmt_list_stmt() -> Stmt {
-    Stmt::Expr(ExprStmt {
-        span: DUMMY_SP,
-        expr: Box::new(Expr::Ident(ident_node(STMT_LIST_HOLE_KEYWORD))),
-    })
-}
-
-fn object_props_prop() -> PropOrSpread {
-    PropOrSpread::Prop(Box::new(Prop::Shorthand(ident_node(
-        OBJECT_PROPS_HOLE_KEYWORD,
-    ))))
-}
-
-/// An `ARGS` argument-list hole that absorbs a run of dropped (non-anchor)
-/// call/`new` arguments (the argument analog of `OBJECT_PROPS`).
-fn args_hole() -> ExprOrSpread {
-    ExprOrSpread {
-        spread: None,
-        expr: Box::new(Expr::Ident(ident_node(ARGS_HOLE_KEYWORD))),
-    }
-}
-
-/// A `case CASE_REST:` switch-case hole that absorbs a run of dropped
-/// `case`/`default` clauses (the switch analog of `CLASS_REST;`).
-fn case_rest_case() -> SwitchCase {
-    SwitchCase {
-        span: DUMMY_SP,
-        test: Some(Box::new(Expr::Ident(ident_node(CASE_REST_HOLE_KEYWORD)))),
-        cons: vec![],
-    }
-}
-
-fn anything_pat() -> Pat {
-    Pat::Ident(BindingIdent {
-        id: ident_node(ANYTHING_HOLE_KEYWORD),
-        type_ann: None,
-    })
-}
-
-fn holed_block(block: &BlockStmt, kept: &BTreeSet<AnchorSpan>) -> BlockStmt {
-    let mut holed = block.clone();
-    holed.stmts = hole_stmts(&block.stmts, kept);
-    holed
-}
-
-/// Hole a function for selector form: every parameter to an `ANYTHING` pattern
-/// (pinning arity, not names) and the body's statements to `STMT_LIST` runs
-/// around the kept anchors. Used both for top-level function selectors and for
-/// function-valued subexpressions reached through [`hole_expr`].
-fn hole_function(function: &Function, kept: &BTreeSet<AnchorSpan>) -> Function {
-    let mut holed = function.clone();
-    holed.params = function.params.iter().map(|_| anything_param()).collect();
-    if let Some(body) = &function.body {
-        holed.body = Some(holed_block(body, kept));
-    }
-    holed
-}
-
-/// Prune an expression into selector form: keep concrete tokens whose span is in
-/// `kept`, and replace every subtree off a kept token's path with an `ANYTHING`
-/// hole node. The result is an ordinary `swc` AST emitted by codegen, not a
-/// hand-built string.
-fn hole_expr(expr: &Expr, kept: &BTreeSet<AnchorSpan>) -> Expr {
-    if !node_retains_any(expr.span(), kept) {
-        return anything_expr();
-    }
-    match expr {
-        Expr::Paren(paren) => hole_expr(&paren.expr, kept),
-        Expr::Lit(_) | Expr::Ident(_) | Expr::Tpl(_) => expr.clone(),
-        Expr::Member(member) => {
-            let mut holed = member.clone();
-            holed.obj = Box::new(hole_expr(&member.obj, kept));
-            Expr::Member(holed)
-        }
-        Expr::Call(call) => {
-            let mut holed = call.clone();
-            holed.callee = hole_callee(&call.callee, kept);
-            holed.args = hole_args(&call.args, kept);
-            Expr::Call(holed)
-        }
-        Expr::New(new_expr) => {
-            let mut holed = new_expr.clone();
-            holed.callee = Box::new(hole_callee_expr(&new_expr.callee, kept));
-            holed.args = new_expr.args.as_ref().map(|args| hole_args(args, kept));
-            Expr::New(holed)
-        }
-        Expr::Object(object) => Expr::Object(hole_object(object, kept)),
-        Expr::Array(array) => Expr::Array(hole_array(array, kept)),
-        Expr::Await(await_expr) => {
-            let mut holed = await_expr.clone();
-            holed.arg = Box::new(hole_expr(&await_expr.arg, kept));
-            Expr::Await(holed)
-        }
-        Expr::Unary(unary) => {
-            let mut holed = unary.clone();
-            holed.arg = Box::new(hole_expr(&unary.arg, kept));
-            Expr::Unary(holed)
-        }
-        Expr::Bin(bin) => {
-            let mut holed = bin.clone();
-            holed.left = Box::new(hole_expr(&bin.left, kept));
-            holed.right = Box::new(hole_expr(&bin.right, kept));
-            Expr::Bin(holed)
-        }
-        // Assignment expression (`state.delta = "value"`): the dominant statement
-        // shape in sequential write-block bodies. Hole the LHS target's receiver
-        // (`state` → `ANYTHING`) while keeping the stable property name, and recurse
-        // into the RHS so a discriminating literal there is kept and everything else
-        // holed. The member property is preserved verbatim, mirroring `Expr::Member`.
-        Expr::Assign(assign) => {
-            let mut holed = assign.clone();
-            holed.left = hole_assign_target(&assign.left, kept);
-            holed.right = Box::new(hole_expr(&assign.right, kept));
-            Expr::Assign(holed)
-        }
-        // Function/arrow-valued subexpressions (e.g. a `wrap(function(){…})` or
-        // `useCallback((e) => {…})` initializer) carrying a kept anchor: hole the
-        // params to `ANYTHING` and the body to `STMT_LIST` around the anchor, the
-        // same interior holing the top-level function selector does. A callback
-        // with no kept anchor never reaches here — the leading `node_retains_any`
-        // guard already collapsed it to `ANYTHING`.
-        Expr::Fn(fn_expr) => {
-            let mut holed = fn_expr.clone();
-            holed.function = Box::new(hole_function(&fn_expr.function, kept));
-            Expr::Fn(holed)
-        }
-        Expr::Arrow(arrow) => {
-            let mut holed = arrow.clone();
-            holed.params = arrow.params.iter().map(|_| anything_pat()).collect();
-            holed.body = Box::new(match arrow.body.as_ref() {
-                BlockStmtOrExpr::BlockStmt(block) => {
-                    BlockStmtOrExpr::BlockStmt(holed_block(block, kept))
-                }
-                BlockStmtOrExpr::Expr(expr) => {
-                    BlockStmtOrExpr::Expr(Box::new(hole_expr(expr, kept)))
-                }
-            });
-            Expr::Arrow(holed)
-        }
-        // Unmodeled shapes carrying a kept anchor: keep verbatim rather than
-        // risk an unsound hole. Over-pinning here is a future refinement.
-        _ => expr.clone(),
-    }
-}
-
-/// Hole an assignment target. A member target (`receiver.prop`) holes its
-/// receiver expression (a minified binding such as a function parameter) while
-/// keeping the stable property name, mirroring [`hole_expr`]'s `Expr::Member`
-/// arm. Bare-ident and pattern targets are kept verbatim — an ident assign
-/// target is itself a (usually minified) name with nothing to hole, and
-/// destructuring patterns are left intact rather than risk an unsound hole.
-fn hole_assign_target(target: &AssignTarget, kept: &BTreeSet<AnchorSpan>) -> AssignTarget {
-    match target {
-        AssignTarget::Simple(SimpleAssignTarget::Member(member)) => {
-            let mut holed = member.clone();
-            holed.obj = Box::new(hole_expr(&member.obj, kept));
-            AssignTarget::Simple(SimpleAssignTarget::Member(holed))
-        }
-        _ => target.clone(),
-    }
-}
-
-/// Hole a call's callee, keeping only an alpha-stable invoked identity. A
-/// member-method name (`.then`, `.bar`) is a stable pin and stays; a
-/// bare-function reference (`make`, `wrapOuter`) is a minified name that churns
-/// every rebuild — the matcher alpha-wildcards it, so it discriminates nothing
-/// and the shape index never proposes it as an anchor — so it holes to
-/// `ANYTHING`.
-fn hole_callee(callee: &Callee, kept: &BTreeSet<AnchorSpan>) -> Callee {
-    match callee {
-        Callee::Expr(expr) => Callee::Expr(Box::new(hole_callee_expr(expr, kept))),
-        Callee::Super(_) | Callee::Import(_) => callee.clone(),
-    }
-}
-
-fn hole_callee_expr(expr: &Expr, kept: &BTreeSet<AnchorSpan>) -> Expr {
-    match expr {
-        // Member-method callee (`a.then(...)`): keep the stable property name even
-        // when no anchor lands on it, holing only the receiver. Routing through
-        // `hole_expr` would instead collapse the whole member to `ANYTHING` when its
-        // subtree carries no kept anchor, dropping the discriminating method name.
-        Expr::Member(member) => {
-            let mut holed = member.clone();
-            holed.obj = Box::new(hole_expr(&member.obj, kept));
-            Expr::Member(holed)
-        }
-        Expr::Paren(paren) => hole_callee_expr(&paren.expr, kept),
-        // Bare-identifier callee (and any other callee expression): hole through the
-        // normal path. A bare-function name is alpha-wildcarded by the matcher and is
-        // never a chosen anchor, so `hole_expr` holes it to `ANYTHING`.
-        _ => hole_expr(expr, kept),
-    }
-}
-
-/// Hole a call/`new` argument list for selector form: a run of dropped
-/// (no-anchor) arguments collapses to a single `ARGS` run hole, while each
-/// argument carrying an anchor is kept and recursively holed.
-///
-/// The `ARGS` hole matches as an ordered subsequence with gaps
-/// (`match_expr_or_spread_slice`), so the selector survives a rebuild that adds
-/// or removes a non-anchor argument — unlike a per-argument `ANYTHING`, which is
-/// an arity-exact single-node hole. Mirrors [`hole_object`]'s `OBJECT_PROPS`
-/// run-collapsing. An anchored spread is kept verbatim (its expr left intact);
-/// a non-anchor spread is absorbed into the run like any other dropped argument.
-fn hole_args(args: &[ExprOrSpread], kept: &BTreeSet<AnchorSpan>) -> Vec<ExprOrSpread> {
-    let mut holed = Vec::new();
-    let mut dropped_run = false;
-    for arg in args {
-        if node_retains_any(arg.expr.span(), kept) {
-            if dropped_run {
-                holed.push(args_hole());
-                dropped_run = false;
-            }
-            let mut kept_arg = arg.clone();
-            if arg.spread.is_none() {
-                kept_arg.expr = Box::new(hole_expr(&arg.expr, kept));
-            }
-            holed.push(kept_arg);
-        } else {
-            dropped_run = true;
-        }
-    }
-    if dropped_run {
-        holed.push(args_hole());
-    }
-    holed
-}
-
-/// Prune the elements of an array literal, holing each non-anchor element to
-/// `ANYTHING` (an arity-exact `EXPR` hole) and recursing into the ones that
-/// carry an anchor. The matcher matches array elements element-wise (there is no
-/// array list hole, unlike `ARGS` for call arguments), so the holed array keeps
-/// the same length; elisions and spreads are preserved verbatim.
-fn hole_array(array: &ArrayLit, kept: &BTreeSet<AnchorSpan>) -> ArrayLit {
-    let mut holed = array.clone();
-    holed.elems = array
-        .elems
-        .iter()
-        .map(|elem| {
-            elem.as_ref().map(|element| {
-                let mut holed_element = element.clone();
-                if element.spread.is_none() {
-                    holed_element.expr = Box::new(hole_expr(&element.expr, kept));
-                }
-                holed_element
-            })
-        })
-        .collect();
-    holed
-}
-
-/// The `OBJECT_PROPS` destructure-pattern hole prop — a shorthand binding whose
-/// name is the keyword — absorbing a run of dropped destructured properties.
-/// The pattern analog of [`object_props_prop`].
-fn object_props_pat_prop() -> ObjectPatProp {
-    ObjectPatProp::Assign(AssignPatProp {
-        span: DUMMY_SP,
-        key: BindingIdent {
-            id: ident_node(OBJECT_PROPS_HOLE_KEYWORD),
-            type_ann: None,
-        },
-        value: None,
-    })
-}
-
-/// Prune a binding pattern into selector form. An object destructuring pattern
-/// (`const { …, x } = e`) holes down to the props carrying a kept anchor (a
-/// discriminating destructured key) via [`hole_object_pat`]; every other
-/// pattern is kept verbatim — a bare minified binding name is already
-/// alpha-wildcarded by the matcher, so there is nothing to hole.
-fn hole_pat(pat: &Pat, kept: &BTreeSet<AnchorSpan>) -> Pat {
-    match pat {
-        Pat::Object(object) => Pat::Object(hole_object_pat(object, kept)),
-        _ => pat.clone(),
-    }
-}
-
-/// Hole a destructuring object pattern: keep only the props carrying a kept
-/// anchor (the discriminating destructured key), dropping every other run into
-/// an `OBJECT_PROPS` hole. The pattern analog of [`hole_object`]; a kept prop is
-/// retained verbatim (its bound local is alpha-wildcarded by the matcher).
-fn hole_object_pat(object: &ObjectPat, kept: &BTreeSet<AnchorSpan>) -> ObjectPat {
-    let mut props = Vec::new();
-    let mut dropped_run = false;
-    for prop in &object.props {
-        if node_retains_any(prop.span(), kept) {
-            if dropped_run {
-                props.push(object_props_pat_prop());
-                dropped_run = false;
-            }
-            props.push(prop.clone());
-        } else {
-            dropped_run = true;
-        }
-    }
-    if dropped_run {
-        props.push(object_props_pat_prop());
-    }
-    let mut holed = object.clone();
-    holed.props = props;
-    holed
-}
-
-fn hole_object(object: &ObjectLit, kept: &BTreeSet<AnchorSpan>) -> ObjectLit {
-    let mut props = Vec::new();
-    let mut dropped_run = false;
-    for prop in &object.props {
-        if node_retains_any(prop.span(), kept) {
-            if dropped_run {
-                props.push(object_props_prop());
-                dropped_run = false;
-            }
-            props.push(hole_prop(prop, kept));
-        } else {
-            dropped_run = true;
-        }
-    }
-    if dropped_run {
-        props.push(object_props_prop());
-    }
-    let mut holed = object.clone();
-    holed.props = props;
-    holed
-}
-
-/// Hole an object literal for the read-off object form: keep only props carrying
-/// a kept anchor, with an `OBJECT_PROPS` run hole before the first kept prop,
-/// after the last, and **between every pair** of kept props.
-///
-/// Unlike [`hole_object`] (which only emits a list hole where a run of props was
-/// actually dropped), this pads both edges and interleaves unconditionally.
-/// Object properties are unordered enum/lookup entries: a kept key can move on a
-/// rebuild, so anchoring one to the object's edge (`anchored_right` in the
-/// matcher) or assuming two kept keys stay adjacent is fragile. Surrounding every
-/// kept prop with `OBJECT_PROPS` matches each as an independent interior
-/// subsequence element, so a minimal *key set* survives key reorder and arbitrary
-/// gaps. With no kept prop this is a bare `{ OBJECT_PROPS }`; with one it is the
-/// padded single-key form (unchanged from the edge-padding behavior).
-fn hole_object_padded(object: &ObjectLit, kept: &BTreeSet<AnchorSpan>) -> ObjectLit {
-    let mut props = vec![object_props_prop()];
-    for prop in &object.props {
-        if node_retains_any(prop.span(), kept) {
-            props.push(hole_prop(prop, kept));
-            props.push(object_props_prop());
-        }
-    }
-    let mut holed = object.clone();
-    holed.props = props;
-    holed
-}
-
-fn hole_prop(prop: &PropOrSpread, kept: &BTreeSet<AnchorSpan>) -> PropOrSpread {
-    let PropOrSpread::Prop(inner) = prop else {
-        return prop.clone();
-    };
-    if let Prop::KeyValue(key_value) = inner.as_ref() {
-        let mut holed = key_value.clone();
-        holed.value = Box::new(hole_expr(&key_value.value, kept));
-        PropOrSpread::Prop(Box::new(Prop::KeyValue(holed)))
-    } else {
-        prop.clone()
-    }
-}
-
-/// Hole a statement list, collapsing runs of dropped statements into a single
-/// `STMT_LIST;` hole statement.
-fn hole_stmts(stmts: &[Stmt], kept: &BTreeSet<AnchorSpan>) -> Vec<Stmt> {
-    let mut out = Vec::new();
-    let mut dropped_run = false;
-    for stmt in stmts {
-        if node_retains_any(stmt.span(), kept) {
-            if dropped_run {
-                out.push(stmt_list_stmt());
-                dropped_run = false;
-            }
-            out.push(hole_stmt(stmt, kept));
-        } else {
-            dropped_run = true;
-        }
-    }
-    if dropped_run || out.is_empty() {
-        out.push(stmt_list_stmt());
-    }
-    out
-}
-
-fn hole_stmt(stmt: &Stmt, kept: &BTreeSet<AnchorSpan>) -> Stmt {
-    match stmt {
-        Stmt::Expr(expr_stmt) => {
-            let mut holed = expr_stmt.clone();
-            holed.expr = Box::new(hole_expr(&expr_stmt.expr, kept));
-            Stmt::Expr(holed)
-        }
-        Stmt::Return(ret) => {
-            let mut holed = ret.clone();
-            holed.arg = ret.arg.as_ref().map(|arg| Box::new(hole_expr(arg, kept)));
-            Stmt::Return(holed)
-        }
-        Stmt::Throw(throw) => {
-            let mut holed = throw.clone();
-            holed.arg = Box::new(hole_expr(&throw.arg, kept));
-            Stmt::Throw(holed)
-        }
-        Stmt::Decl(Decl::Var(var)) => {
-            let mut holed = (**var).clone();
-            for declarator in &mut holed.decls {
-                declarator.name = hole_pat(&declarator.name, kept);
-                if let Some(init) = &declarator.init {
-                    declarator.init = Some(Box::new(hole_expr(init, kept)));
-                }
-            }
-            Stmt::Decl(Decl::Var(Box::new(holed)))
-        }
-        Stmt::If(if_stmt) => {
-            let mut holed = if_stmt.clone();
-            holed.test = Box::new(hole_expr(&if_stmt.test, kept));
-            let cons_stmts = match if_stmt.cons.as_ref() {
-                Stmt::Block(block) => hole_stmts(&block.stmts, kept),
-                other => hole_stmts(std::slice::from_ref(other), kept),
-            };
-            holed.cons = Box::new(Stmt::Block(BlockStmt {
-                span: DUMMY_SP,
-                ctxt: SyntaxContext::empty(),
-                stmts: cons_stmts,
-            }));
-            holed.alt = None;
-            Stmt::If(holed)
-        }
-        Stmt::Try(try_stmt) => {
-            let mut holed = try_stmt.clone();
-            holed.block = holed_block(&try_stmt.block, kept);
-            if let Some(handler) = &mut holed.handler {
-                if handler.param.is_some() {
-                    handler.param = Some(anything_pat());
-                }
-                handler.body = holed_block(&handler.body, kept);
-            }
-            if let Some(finalizer) = &mut holed.finalizer {
-                *finalizer = holed_block(finalizer, kept);
-            }
-            Stmt::Try(holed)
-        }
-        Stmt::Block(block) => Stmt::Block(holed_block(block, kept)),
-        Stmt::Switch(switch) => {
-            let mut holed = switch.clone();
-            holed.discriminant = Box::new(hole_expr(&switch.discriminant, kept));
-            holed.cases = hole_switch_cases(&switch.cases, kept);
-            Stmt::Switch(holed)
-        }
-        // Unmodeled statement shapes carrying a kept anchor: keep verbatim.
-        _ => stmt.clone(),
-    }
-}
-
-/// Prune a `switch`'s case list: drop runs of non-discriminating
-/// `case`/`default` clauses into `case CASE_REST:` holes, keeping only the
-/// clauses that retain an anchor (their test literal or a body statement).
-/// Mirrors [`hole_class_members`].
-fn hole_switch_cases(cases: &[SwitchCase], kept: &BTreeSet<AnchorSpan>) -> Vec<SwitchCase> {
-    let mut out = Vec::new();
-    let mut dropped_run = false;
-    for case in cases {
-        if node_retains_any(case.span(), kept) {
-            if dropped_run {
-                out.push(case_rest_case());
-                dropped_run = false;
-            }
-            out.push(hole_switch_case(case, kept));
-        } else {
-            dropped_run = true;
-        }
-    }
-    if dropped_run || out.is_empty() {
-        out.push(case_rest_case());
-    }
-    out
-}
-
-fn hole_switch_case(case: &SwitchCase, kept: &BTreeSet<AnchorSpan>) -> SwitchCase {
-    let mut holed = case.clone();
-    holed.test = case
-        .test
-        .as_ref()
-        .map(|test| Box::new(hole_expr(test, kept)));
-    holed.cons = hole_stmts(&case.cons, kept);
-    holed
-}
-
-/// Candidate concrete anchors, split into preference tiers. Literal values are
-/// stable, meaningful landmarks (a magic string, a config number/bool), so they
-/// are tried before structural member/property and key names. Among literals,
-/// shallower ones (fewer enclosing call/new levels) are preferred: a direct
-/// `key: "primary"` pins less nested structure — and is more rebuild-robust —
-/// than a literal buried inside `mk("primary")`.
-#[derive(Default)]
-struct AnchorCandidates {
-    /// `(span, call-nesting depth)` for each literal.
-    literals: Vec<(AnchorSpan, u32)>,
-    structural: Vec<AnchorSpan>,
-}
-
-/// A literal at most this many call/new levels deep counts as a *direct* value
-/// or argument — a meaningful, stable pin worth keeping. Deeper literals are
-/// treated as buried in incidental computation and ranked below structural
-/// key/member presence.
-const SHALLOW_LITERAL_DEPTH: u32 = 1;
-
-impl AnchorCandidates {
-    /// Shallow literal anchors — direct values/args worth keeping as meaningful
-    /// per-slot pins even when structure alone would already discriminate. This
-    /// includes object-property values (`kind: "primary"`): the unified anchor
-    /// policy favors keeping shallow literals over an exact-minimum cover, so an
-    /// occasional over-pin (a shared `enabled: true`) is accepted as the price
-    /// of a single policy across single and group targets.
-    fn shallow_literals(&self) -> Vec<AnchorSpan> {
-        self.literals
-            .iter()
-            .filter(|(_, depth)| *depth <= SHALLOW_LITERAL_DEPTH)
-            .map(|(span, _)| *span)
-            .collect()
-    }
-
-    /// Tiers consulted only when shallow literals are not enough on their own:
-    /// structural key/member presence, then deeper literals by ascending depth.
-    fn deep_cover_tiers(&self) -> Vec<Vec<AnchorSpan>> {
-        let mut tiers = Vec::new();
-        if !self.structural.is_empty() {
-            tiers.push(self.structural.clone());
-        }
-        if let Some(max_depth) = self.literals.iter().map(|(_, depth)| *depth).max() {
-            for depth in (SHALLOW_LITERAL_DEPTH + 1)..=max_depth {
-                let tier: Vec<AnchorSpan> = self
-                    .literals
-                    .iter()
-                    .filter(|(_, anchor_depth)| *anchor_depth == depth)
-                    .map(|(span, _)| *span)
-                    .collect();
-                if !tier.is_empty() {
-                    tiers.push(tier);
-                }
-            }
-        }
-        tiers
-    }
-}
-
-/// `depth` counts enclosing call/new levels, so the keep-shallow group path can
-/// prefer shallower literal anchors. Object-property values (`kind: "primary"`)
-/// are collected at the enclosing depth like any other literal; the keep-shallow
-/// policy retains them rather than treating them as incidental.
-fn collect_expr_anchors(expr: &Expr, depth: u32, candidates: &mut AnchorCandidates) {
-    match expr {
-        Expr::Lit(_) | Expr::Tpl(_) => {
-            candidates.literals.push((span_key(expr.span()), depth));
-        }
-        Expr::Paren(paren) => collect_expr_anchors(&paren.expr, depth, candidates),
-        Expr::Member(member) => {
-            collect_expr_anchors(&member.obj, depth, candidates);
-            if let MemberProp::Ident(ident) = &member.prop {
-                candidates.structural.push(span_key(ident.span));
-            } else if let MemberProp::Computed(computed) = &member.prop {
-                collect_expr_anchors(&computed.expr, depth, candidates);
-            }
-        }
-        Expr::Call(call) => {
-            if let Callee::Expr(callee) = &call.callee {
-                collect_expr_anchors(callee, depth, candidates);
-            }
-            for arg in &call.args {
-                if arg.spread.is_none() {
-                    collect_expr_anchors(&arg.expr, depth + 1, candidates);
-                }
-            }
-        }
-        Expr::New(new_expr) => {
-            collect_expr_anchors(&new_expr.callee, depth, candidates);
-            for arg in new_expr.args.as_deref().unwrap_or_default() {
-                if arg.spread.is_none() {
-                    collect_expr_anchors(&arg.expr, depth + 1, candidates);
-                }
-            }
-        }
-        Expr::Object(object) => {
-            for prop in &object.props {
-                if let PropOrSpread::Prop(prop) = prop {
-                    if let Prop::KeyValue(key_value) = prop.as_ref() {
-                        candidates.structural.push(span_key(key_value.key.span()));
-                        collect_expr_anchors(&key_value.value, depth, candidates);
-                    }
-                }
-            }
-        }
-        Expr::Await(await_expr) => collect_expr_anchors(&await_expr.arg, depth, candidates),
-        Expr::Unary(unary) => collect_expr_anchors(&unary.arg, depth, candidates),
-        Expr::Bin(bin) => {
-            collect_expr_anchors(&bin.left, depth, candidates);
-            collect_expr_anchors(&bin.right, depth, candidates);
-        }
-        _ => {}
-    }
-}
-
 /// Every `(body, binding-slot)` alignment a `target_binding` selector with
 /// `match_source` resolves to.
 ///
@@ -2205,1229 +1557,6 @@ fn matched_body_indices(
             .map(|candidate| candidate.body_idx)
             .collect(),
     )
-}
-
-fn holes_present(source: &str) -> BTreeSet<String> {
-    let mut holes = BTreeSet::new();
-    for keyword in [
-        ANYTHING_HOLE_KEYWORD,
-        STMT_LIST_HOLE_KEYWORD,
-        OBJECT_PROPS_HOLE_KEYWORD,
-        CLASS_REST_HOLE_KEYWORD,
-        DECLARATORS_HOLE_KEYWORD,
-    ] {
-        if source.contains(keyword) {
-            holes.insert(keyword.to_string());
-        }
-    }
-    holes
-}
-
-fn minimize_function_selector(
-    index: &ChunkSelectorIndex,
-    function: &Function,
-    decl: &IndexedDeclaration,
-    target: &SynthesizedTargetBinding,
-) -> Result<Option<SpecializedSelector>> {
-    if function.body.is_none() {
-        return Ok(None);
-    }
-    let render_with = |kept: &BTreeSet<AnchorSpan>| -> Result<String> {
-        emit_selector(ModuleItem::Stmt(Stmt::Decl(Decl::Fn(FnDecl {
-            ident: ident_node(&target.export_name),
-            declare: false,
-            function: Box::new(hole_function(function, kept)),
-        }))))
-    };
-    // Single-target functions read their minimal anchor set off the shape index.
-    // The holed scaffold already pins the param arity (one `ANYTHING` per real
-    // param), so a structural / skeleton anchor needs no kept span; value anchors
-    // (string/number/bool literals, member/method names, member-path callees) map
-    // to the spans of the tokens that exhibit them. The matcher proves the result
-    // (gate 1); a target the read-off cannot single out returns `None` and is
-    // reported as debt (never a full-AST pin).
-    render_via_read_off(index, decl, target, &render_with)
-}
-
-/// Render a single-target selector via the read-off API.
-///
-/// Reads the minimal [`AnchorSet`] off the shape index, maps its value anchors
-/// to kept byte spans over the target item, and renders + proves through the
-/// supplied `render_with` (the same prune + codegen — no second serializer).
-///
-/// **Robustness-anchor policy.** A holed-down *value* anchor is preferred over
-/// the bare structural scaffold even when the scaffold alone would resolve. A
-/// scaffold that pins only declaration kind + arity (`class X { CLASS_REST }`,
-/// `function f(ANYTHING) { STMT_LIST }`, `const X = ANYTHING`) is a degenerate
-/// selector: it pins nothing rebuild-stable and matches any same-shape sibling a
-/// rebuild adds. So the read-off keeps its chosen value anchor when it has one,
-/// and falls back to the bare scaffold only when the target has no renderable
-/// value anchor — `minimal_anchor_set` chose a purely structural skeleton (empty
-/// kept spans) because nothing but shape discriminates — or the value anchor
-/// fails to prove. Returns `None` when neither route singles the target out (a
-/// genuine alpha-duplicate); the caller reports it as debt, never a full-AST pin.
-fn render_via_read_off(
-    index: &ChunkSelectorIndex,
-    decl: &IndexedDeclaration,
-    target: &SynthesizedTargetBinding,
-    render_with: &impl Fn(&BTreeSet<AnchorSpan>) -> Result<String>,
-) -> Result<Option<SpecializedSelector>> {
-    let item = index
-        .parsed
-        .module
-        .body
-        .get(decl.body_idx)
-        .context("read-off body index no longer in module")?;
-
-    // Primary read-off: the single minimal anchor set. When its value anchor
-    // renders to a holed selector that proves uniquely, keep it.
-    if let Some(anchor_set) = index.shape_index.minimal_anchor_set(decl.body_idx) {
-        let kept = kept_spans_for_anchor_set(item, &anchor_set);
-        if !kept.is_empty()
-            && let Some(selector) =
-                finish_minimized_selector(index, decl, target, render_with(&kept)?)?
-        {
-            return Ok(Some(selector));
-        }
-    }
-
-    // Robustness-anchor fallback: the minimal anchor's *value* may not survive
-    // holing — a deep literal whose only occurrence sits inside a large statement
-    // the holer keeps verbatim leaves raw subtrees the matcher rejects, and a
-    // structural-only minimal set keeps nothing at all. Rather than collapse to
-    // the degenerate scaffold, walk the target's individually-discriminating value
-    // anchors best-first and emit the first whose holed selector proves uniquely.
-    // This is what drills a whole-body class/component down to one anchored member
-    // (e.g. `applyChange(ANYTHING) { STMT_LIST; ANYTHING.set("running"); }`,
-    // anchoring on the `"running"` literal) instead of `class X { CLASS_REST }`.
-    for anchor_set in index
-        .shape_index
-        .unique_value_anchor_candidates(decl.body_idx)
-    {
-        let kept = kept_spans_for_anchor_set(item, &anchor_set);
-        if !kept.is_empty()
-            && let Some(selector) =
-                finish_minimized_selector(index, decl, target, render_with(&kept)?)?
-        {
-            return Ok(Some(selector));
-        }
-    }
-
-    // Multi-feature interior cover (#2289): no single value anchor singled the
-    // target out (every value anchor still shares same-shape siblings), but a
-    // *combination* of value anchors may — each individually shared with a
-    // different sibling. Greedy-cover over value-bearing features only, so each
-    // chosen anchor still renders a kept span, and emit the holed selector if it
-    // proves uniquely. This drills bodies whose discriminator is a *set* of deep
-    // leaves rather than one.
-    if let Some(anchor_set) = index.shape_index.unique_value_anchor_cover(decl.body_idx) {
-        let kept = kept_spans_for_anchor_set(item, &anchor_set);
-        if !kept.is_empty()
-            && let Some(selector) =
-                finish_minimized_selector(index, decl, target, render_with(&kept)?)?
-        {
-            return Ok(Some(selector));
-        }
-    }
-
-    // Last resort: the bare structural scaffold, used only when it resolves
-    // uniquely (a purely structural discriminator — arity/shape — with no value
-    // anchor to keep). The scaffold is degenerate for `var`, so the var path skips
-    // this entirely and returns `None` for the keep-shallow path to handle.
-    let empty = BTreeSet::new();
-    if matched_body_indices(index, &target.export_name, &render_with(&empty)?)?
-        == BTreeSet::from([decl.body_idx])
-    {
-        return finish_minimized_selector(index, decl, target, render_with(&empty)?);
-    }
-    Ok(None)
-}
-
-// ===========================================================================
-// Regex-over-string-literal anchors (`STR_LITERAL_MATCHING_RE("<pattern>")`).
-//
-// A var-binding selector that pins an exact string literal breaks whenever a
-// rebuild perturbs a volatile fragment of that literal (a content hash, build
-// counter, or other generated tail). When the *stable* part of the literal
-// already discriminates the target from its siblings, we can pin that stable
-// structure with a regex and wildcard the volatile fragment, so the selector
-// survives the rebuild.
-//
-// Derivation rule (intentionally conservative — see
-// `selector_minimizer_discrimination.md`):
-//
-//   * We only derive a pattern when the literal ends in a *volatile tail*: a
-//     trailing run of hex/digits. The tail must be at least
-//     `MIN_VOLATILE_TAIL_LEN` chars so a one- to three-char numeric suffix
-//     (which is more likely meaningful than generated) is left alone. Any
-//     separator before the tail (`chunk-`, `main.`) stays in the pinned prefix.
-//   * The derived pattern is `^<escaped stable prefix><tail class>$`, anchored
-//     so `Regex::is_match` (which is otherwise a substring test) pins the whole
-//     value. The stable prefix is escaped with `regex::escape`, so every
-//     metacharacter in the literal is matched literally; only the volatile tail
-//     becomes a character-class wildcard (`[0-9A-Fa-f]+` for a hex tail,
-//     `[0-9]+` for a pure-digit tail).
-//   * If the whole literal would be the volatile tail (no stable prefix), we
-//     return `None`: a bare `^[0-9]+$` pins nothing meaningful and would almost
-//     never discriminate.
-//
-// Limits we deliberately accept: this recognizes only trailing hex/digit
-// volatility (the dominant bundler pattern: `chunk-a1b2c3`, `main.4f3a2b.js`,
-// `vendor_1024`). It does not model embedded volatile fragments, GUID shapes,
-// or base64 hashes. The cover never *requires* a regex anchor — it is offered
-// only as an upgrade of an already-kept exact literal, and is taken only when
-// the upgraded selector still resolves uniquely (so an over-broad pattern is
-// rejected, never emitted). A pattern that fails to compile as `regex::Regex`
-// is likewise never emitted.
-// ===========================================================================
-
-/// Minimum length of a trailing hex/digit run for it to count as a volatile
-/// fragment worth wildcarding. Short numeric suffixes (`v2`, `s3`) are more
-/// often meaningful than generated, so they stay pinned exactly.
-const MIN_VOLATILE_TAIL_LEN: usize = 4;
-
-/// Derive an anchored `STR_LITERAL_MATCHING_RE` pattern that pins the stable
-/// prefix of `value` and wildcards a trailing volatile hex/digit fragment, or
-/// `None` when no meaningful stable-prefix/volatile-tail split exists. The
-/// returned pattern is always a valid `regex::Regex` (the only metacharacters
-/// it introduces are the anchors and a character class; the prefix is escaped).
-fn regex_anchor_pattern(value: &str) -> Option<String> {
-    let chars: Vec<char> = value.chars().collect();
-    // Length of the trailing run of hex digits.
-    let hex_tail = chars
-        .iter()
-        .rev()
-        .take_while(|c| c.is_ascii_hexdigit())
-        .count();
-    if hex_tail < MIN_VOLATILE_TAIL_LEN {
-        return None;
-    }
-    // Prefer a pure-digit tail class when the whole tail is decimal; otherwise
-    // a hex class. (Hex is a superset of digits, so the digit class is the
-    // tighter, more honest wildcard when applicable.)
-    let tail_is_decimal = chars[chars.len() - hex_tail..]
-        .iter()
-        .all(char::is_ascii_digit);
-    let tail_class = if tail_is_decimal {
-        "[0-9]+"
-    } else {
-        "[0-9A-Fa-f]+"
-    };
-    let stable_prefix: String = chars[..chars.len() - hex_tail].iter().collect();
-    // A regex anchor must pin *something* stable: an empty or separator-only
-    // prefix discriminates nothing, so decline. The trailing separator (if any)
-    // stays in the pinned, escaped prefix — `chunk-a1b2c3` pins `chunk-`, not
-    // `chunk`, which is the conservative choice (one fewer wildcarded char).
-    if stable_prefix.is_empty() || stable_prefix.chars().all(|c| matches!(c, '-' | '_' | '.')) {
-        return None;
-    }
-    let pattern = format!("^{}{tail_class}$", regex::escape(&stable_prefix));
-    // Guard the invariant directly: never hand the matcher a pattern it cannot
-    // compile (the matcher silently treats an uncompilable pattern as a
-    // non-match, which would make the selector match nothing).
-    regex::Regex::new(&pattern).ok()?;
-    Some(pattern)
-}
-
-/// Build the `STR_LITERAL_MATCHING_RE("<pattern>")` call expression the matcher
-/// interprets as a regex-over-string-literal predicate.
-fn regex_predicate_call(pattern: &str) -> Expr {
-    Expr::Call(CallExpr {
-        span: DUMMY_SP,
-        ctxt: SyntaxContext::empty(),
-        callee: Callee::Expr(Box::new(Expr::Ident(ident_node(
-            STRING_LITERAL_REGEX_PREDICATE,
-        )))),
-        args: vec![ExprOrSpread {
-            spread: None,
-            expr: Box::new(Expr::Lit(Lit::Str(Str {
-                span: DUMMY_SP,
-                value: pattern.into(),
-                raw: None,
-            }))),
-        }],
-        type_args: None,
-    })
-}
-
-/// Replace each kept string literal whose span is a chosen regex anchor with the
-/// `STR_LITERAL_MATCHING_RE` predicate call. Runs as a post-pass over the holed
-/// selector AST: holed literals keep their original source span, so matching by
-/// span here is exact and never touches a literal the cover did not select.
-struct RegexAnchorSubstitution<'a> {
-    patterns: &'a BTreeMap<AnchorSpan, String>,
-}
-
-impl VisitMut for RegexAnchorSubstitution<'_> {
-    fn visit_mut_expr(&mut self, expr: &mut Expr) {
-        if let Expr::Lit(Lit::Str(str_lit)) = expr {
-            if let Some(pattern) = self.patterns.get(&span_key(str_lit.span)) {
-                *expr = regex_predicate_call(pattern);
-                return;
-            }
-        }
-        expr.visit_mut_children_with(self);
-    }
-}
-
-/// Candidate regex anchors for the var-binding minimizer: the `(span, pattern)`
-/// of each string literal in a target slot's init for which `regex_anchor_pattern`
-/// yields a wildcarding pattern. Only literals already in the kept set are ever
-/// upgraded, so this is a superset filtered against `kept` at upgrade time.
-fn collect_regex_anchor_candidates(init: &Expr) -> BTreeMap<AnchorSpan, String> {
-    #[derive(Default)]
-    struct Collector {
-        candidates: BTreeMap<AnchorSpan, String>,
-    }
-    impl Visit for Collector {
-        fn visit_expr(&mut self, expr: &Expr) {
-            if let Expr::Lit(Lit::Str(str_lit)) = expr {
-                if let Some(pattern) =
-                    regex_anchor_pattern(str_lit.value.to_string_lossy().as_ref())
-                {
-                    self.candidates.insert(span_key(str_lit.span), pattern);
-                }
-            }
-            expr.visit_children_with(self);
-        }
-    }
-    let mut collector = Collector::default();
-    init.visit_with(&mut collector);
-    collector.candidates
-}
-
-/// Among kept string literals (`candidates` = span → derivable volatile-tail
-/// pattern), accept each `STR_LITERAL_MATCHING_RE` upgrade iff the upgraded
-/// selector *still* resolves uniquely (gate 1). The exact-literal form is the
-/// default; regex is opt-in by merit. Upgrades are applied one literal at a time
-/// so a too-broad pattern on one literal never blocks a sound upgrade on another.
-/// Shared by the read-off single-target var path and the keep-shallow group path.
-fn accepted_regex_anchors(
-    index: &ChunkSelectorIndex,
-    decl: &IndexedDeclaration,
-    targets: &[SynthesizedTargetBinding],
-    candidates: &BTreeMap<AnchorSpan, String>,
-    kept: &BTreeSet<AnchorSpan>,
-    render_with: &impl Fn(&BTreeSet<AnchorSpan>, &BTreeMap<AnchorSpan, String>) -> Result<String>,
-) -> Result<BTreeMap<AnchorSpan, String>> {
-    let mut regex_anchors: BTreeMap<AnchorSpan, String> = BTreeMap::new();
-    for (&span, pattern) in candidates {
-        if !kept.contains(&span) {
-            continue;
-        }
-        regex_anchors.insert(span, pattern.clone());
-        if prove_synthesized_selector(index, decl, targets, &render_with(kept, &regex_anchors)?)
-            .is_err()
-        {
-            // The regex anchor broke uniqueness (over-broad among siblings), so
-            // back it out and keep the exact literal.
-            regex_anchors.remove(&span);
-        }
-    }
-    Ok(regex_anchors)
-}
-
-/// A `DECLARATORS_<pos> = null` declarator hole absorbing a run of non-target
-/// declarators in a binding-group selector.
-fn declarator_hole(name: &str) -> VarDeclarator {
-    VarDeclarator {
-        span: DUMMY_SP,
-        name: named_pat(name),
-        init: Some(Box::new(Expr::Lit(Lit::Null(Null { span: DUMMY_SP })))),
-        definite: false,
-    }
-}
-
-/// Ranked anchor spans for an object literal's key-set cover, best-first: each
-/// direct literal/template **value** (a value-level discriminator, tried first)
-/// then each **key** (the key-set discriminator). Value member accesses and
-/// nested expressions are intentionally excluded — a minified `.prop` is
-/// rebuild-volatile, so the cover pins the stable key and holes the value to
-/// `ANYTHING` rather than anchoring on a churning property name.
-///
-/// Within each class the order is source order; the cover's slot-resolution
-/// greedy ([`cover_object_slot`]) then picks the most discriminating anchor, so
-/// this ordering only breaks ties — preferring a unique value (`accent:
-/// "…Accent"`) over its equally-unique key (`accent: ANYTHING`) when both single
-/// out the slot.
-fn object_anchor_ranking(object: &ObjectLit) -> Vec<AnchorSpan> {
-    let key_value_props = || {
-        object.props.iter().filter_map(|prop| match prop {
-            PropOrSpread::Prop(prop) => match prop.as_ref() {
-                Prop::KeyValue(key_value) => Some(key_value),
-                _ => None,
-            },
-            PropOrSpread::Spread(_) => None,
-        })
-    };
-    let values = key_value_props()
-        .filter(|kv| matches!(kv.value.as_ref(), Expr::Lit(_) | Expr::Tpl(_)))
-        .map(|kv| span_key(kv.value.span()));
-    let keys = key_value_props().map(|kv| span_key(kv.key.span()));
-    values.chain(keys).collect()
-}
-
-/// Slot-aware minimal cover for a single-target object inside a `var` group: a
-/// greedy key-set set-cover that, at each step, adds the `ranked` anchor that best
-/// steers the selector toward resolving to the target binding's own declarator
-/// slot, until it proves unique. The chunk-wide read-off resolves by distinct
-/// body index and so cannot see two sibling declarators of the *same* statement;
-/// this scores by whether the **target binding** is the one the selector
-/// resolves, then by the match count.
-///
-/// The matcher reports one alignment per body (the leftmost declarator the holed
-/// pattern fits), so a key shared with an *earlier* sibling slot
-/// (`blue: "#00f"`, also in `firstPalette`) resolves there, not to the target —
-/// hence the score's first key is "did the target slot resolve at all", which a
-/// key/value unique to the target slot (`accent`) flips true. Keeps adding the
-/// best anchor until the matcher's uniqueness proof passes, or the target's own
-/// anchors are exhausted (then `None`, and the caller keeps its keep-shallow
-/// form).
-fn cover_object_slot(
-    index: &ChunkSelectorIndex,
-    decl: &IndexedDeclaration,
-    target: &SynthesizedTargetBinding,
-    ranked: &[AnchorSpan],
-    render_with: &impl Fn(&BTreeSet<AnchorSpan>) -> Result<String>,
-) -> Result<Option<SpecializedSelector>> {
-    let targets = std::slice::from_ref(target);
-    let mut kept: BTreeSet<AnchorSpan> = BTreeSet::new();
-    while prove_synthesized_selector(index, decl, targets, &render_with(&kept)?).is_err() {
-        // Score a trial by `(target slot not yet resolved, total matches)`; a
-        // smaller score is better, so an anchor that makes the target binding the
-        // resolved one and rules out the most competitors wins.
-        let mut best: Option<((bool, usize), AnchorSpan)> = None;
-        for &anchor in ranked.iter().take(MAX_MINIMIZER_ANCHORS) {
-            if kept.contains(&anchor) {
-                continue;
-            }
-            let mut trial = kept.clone();
-            trial.insert(anchor);
-            let matches =
-                matched_binding_candidates(index, &target.export_name, &render_with(&trial)?)?;
-            let target_unresolved = !matches.iter().any(|m| {
-                m.body_idx == decl.body_idx && m.binding.binding_name == target.runtime_binding
-            });
-            let score = (target_unresolved, matches.len());
-            if best.is_none_or(|(best_score, _)| score < best_score) {
-                best = Some((score, anchor));
-            }
-        }
-        // No remaining anchor to add: the target's own keys/values cannot single
-        // out its slot. Defer to the caller.
-        let Some((_, anchor)) = best else {
-            return Ok(None);
-        };
-        kept.insert(anchor);
-    }
-    finish_minimized_selector(index, decl, target, render_with(&kept)?)
-}
-
-/// Read off a minimal selector for a single-target `var`/`let`/`const` whose
-/// target declarator value is an object literal (W3 + key-set minimization).
-///
-/// Handles the single-target object whether it stands alone or sits inside a
-/// multi-declarator group (one target slot, `DECLARATORS_*` holes for the rest).
-/// Two passes, both rendering through one slot-aware `render_with` that holes the
-/// object with [`hole_object_padded`] (interleaved `OBJECT_PROPS`) so the kept key
-/// subset survives key reorder:
-///
-///   1. **Read-off.** The chunk-wide shape index ranks the statement's own
-///      features by selective × stable, so a globally-rare discriminating
-///      key/value wins. Its kept spans are restricted to the target declarator (a
-///      group's anchor set may name a key carried only by a *sibling* declarator
-///      this selector holes away), and the matcher proves the restricted pin.
-///   2. **Cover fallback.** A matcher-driven minimal cover over the target
-///      object's own keys (and direct literal values) — the key-set analogue of
-///      greedy set-cover: anchor the rarest discriminating keys, `OBJECT_PROPS`
-///      the common ones. This is what singles out a target object inside a
-///      multi-declarator group, where the chunk-wide read-off cannot see the
-///      per-slot key sets.
-///
-/// Returns `None` — so the caller falls back to the keep-shallow group path —
-/// when the target is not a single object declarator or neither pass resolves it.
-fn try_object_read_off(
-    index: &ChunkSelectorIndex,
-    var: &VarDecl,
-    decl: &IndexedDeclaration,
-    targets: &[SynthesizedTargetBinding],
-    target_slots: &BTreeSet<usize>,
-) -> Result<Option<SpecializedSelector>> {
-    // Only the single-target case (one binding). A multi-target group needs the
-    // tuple-resolving binding-group path; this owns the single object slot.
-    let [target] = targets else {
-        return Ok(None);
-    };
-    if target_slots.len() != 1 {
-        return Ok(None);
-    }
-    let target_slot = *target_slots.iter().next().expect("one target slot");
-    let declarator = &var.decls[target_slot];
-    let Some(Expr::Object(object)) = declarator.init.as_deref() else {
-        return Ok(None);
-    };
-    let target_decl_span = declarator.span();
-
-    // Slot-aware render via the shared var-slot renderer: `DECLARATORS_*` holes
-    // for the non-target declarators (none when the target stands alone), the
-    // target's object holed to its kept keys padded + interleaved with
-    // `OBJECT_PROPS` (`hole_var_init_padded`'s object arm). The object path never
-    // upgrades to a regex anchor, so the regex map is always empty.
-    let only_target = BTreeSet::from([target_slot]);
-    let no_regex: BTreeMap<AnchorSpan, String> = BTreeMap::new();
-    let export_for =
-        |name: &str| (name == target.runtime_binding).then(|| target.export_name.clone());
-    let render_with = |kept: &BTreeSet<AnchorSpan>| -> Result<String> {
-        render_var_slots(
-            var,
-            &only_target,
-            &export_for,
-            kept,
-            &no_regex,
-            &hole_var_init_padded,
-        )
-    };
-
-    if let Some(anchor_set) = index.shape_index.minimal_anchor_set(decl.body_idx) {
-        let item = index
-            .parsed
-            .module
-            .body
-            .get(decl.body_idx)
-            .context("read-off body index no longer in module")?;
-        let kept: BTreeSet<AnchorSpan> = kept_spans_for_anchor_set(item, &anchor_set)
-            .into_iter()
-            .filter(|span| node_holds_anchor(target_decl_span, *span))
-            .collect();
-        if !kept.is_empty()
-            && let Some(selector) =
-                finish_minimized_selector(index, decl, target, render_with(&kept)?)?
-        {
-            return Ok(Some(selector));
-        }
-    }
-
-    cover_object_slot(
-        index,
-        decl,
-        target,
-        &object_anchor_ranking(object),
-        &render_with,
-    )
-}
-
-/// Read off a minimal selector for a single-target `var`/`let`/`const` whose
-/// target declarator value is **not** an object literal (objects route through
-/// [`try_object_read_off`]). Mirrors [`render_via_read_off`] (function/class) but
-/// slot-aware: holes non-target declarators to `DECLARATORS_*`, holes the target
-/// init with [`hole_expr`] around the read-off anchors, and restricts the kept
-/// spans to the target declarator so a group's chunk-wide anchor set never pins a
-/// span carried only by a sibling this selector holes away.
-///
-/// Deviation from the function/class read-off: there is no empty-kept structural
-/// fast path. A var's maximally-holed scaffold is the *degenerate* `const X =
-/// ANYTHING`, which pins nothing meaningful and would match any wrapped-const a
-/// rebuild adds (criterion 2). A var selector must keep a discriminating value
-/// anchor, so an empty anchor set yields `None` and the caller falls back to the
-/// keep-shallow group path.
-///
-/// Returns `None` — caller falls back — when the target is an object declarator,
-/// the read-off has no anchor set, every anchor lies outside the target slot, or
-/// the rendered selector fails the matcher gate.
-fn try_var_read_off(
-    index: &ChunkSelectorIndex,
-    var: &VarDecl,
-    decl: &IndexedDeclaration,
-    target: &SynthesizedTargetBinding,
-    target_slot: usize,
-) -> Result<Option<SpecializedSelector>> {
-    let declarator = &var.decls[target_slot];
-    let Some(init) = declarator.init.as_deref() else {
-        return Ok(None);
-    };
-    // Objects are `try_object_read_off`'s domain (padded `OBJECT_PROPS` + the
-    // slot-aware key-set cover); this owns every other initializer shape.
-    if matches!(init, Expr::Object(_)) {
-        return Ok(None);
-    }
-    let target_decl_span = declarator.span();
-    let targets = std::slice::from_ref(target);
-    let only_target = BTreeSet::from([target_slot]);
-    let export_for =
-        |name: &str| (name == target.runtime_binding).then(|| target.export_name.clone());
-
-    // Shared var-slot render: `DECLARATORS_*` holes for the non-target declarators
-    // (none when the target stands alone), the target's non-object init holed via
-    // `hole_var_init_padded` (the `other` arm, i.e. `hole_expr`).
-    let render_with = |kept: &BTreeSet<AnchorSpan>,
-                       regex_anchors: &BTreeMap<AnchorSpan, String>|
-     -> Result<String> {
-        render_var_slots(
-            var,
-            &only_target,
-            &export_for,
-            kept,
-            regex_anchors,
-            &hole_var_init_padded,
-        )
-    };
-
-    let item = index
-        .parsed
-        .module
-        .body
-        .get(decl.body_idx)
-        .context("read-off body index no longer in module")?;
-    let no_regex = BTreeMap::new();
-
-    // Try the single minimal anchor set first, then — robustness-anchor fallback —
-    // the target's individually-discriminating value anchors best-first. A deep
-    // value anchor whose minimal pick does not prove (it lands in a statement the
-    // holer keeps verbatim, leaving raw subtrees the matcher rejects) is recovered
-    // here instead of collapsing to the degenerate `const X = ANYTHING` scaffold;
-    // this drills a whole-body component initializer down to one anchored leaf.
-    // Finally the multi-feature interior cover (#2289): when no single value anchor
-    // is unique, a greedy cover over value-bearing features keeps the *set* of deep
-    // leaves that jointly single the slot out.
-    let anchor_sets = index
-        .shape_index
-        .minimal_anchor_set(decl.body_idx)
-        .into_iter()
-        .chain(
-            index
-                .shape_index
-                .unique_value_anchor_candidates(decl.body_idx),
-        )
-        .chain(index.shape_index.unique_value_anchor_cover(decl.body_idx));
-    for anchor_set in anchor_sets {
-        let kept: BTreeSet<AnchorSpan> = kept_spans_for_anchor_set(item, &anchor_set)
-            .into_iter()
-            .filter(|span| node_holds_anchor(target_decl_span, *span))
-            .collect();
-        if kept.is_empty() {
-            continue;
-        }
-        // Prove the read-off resolves uniquely before offering the regex upgrade.
-        if prove_synthesized_selector(index, decl, targets, &render_with(&kept, &no_regex)?)
-            .is_err()
-        {
-            continue;
-        }
-        // Preserve the keep-shallow path's regex-literal upgrade: among kept string
-        // literals, swap a volatile-suffix value for a `STR_LITERAL_MATCHING_RE`
-        // anchor when the upgraded selector still resolves uniquely.
-        let regex_anchors = accepted_regex_anchors(
-            index,
-            decl,
-            targets,
-            &collect_regex_anchor_candidates(init),
-            &kept,
-            &render_with,
-        )?;
-        let source = render_with(&kept, &regex_anchors)?;
-        let rewritten_holes = holes_present(&source);
-        return Ok(Some(SpecializedSelector {
-            match_source: source,
-            rewritten_holes,
-        }));
-    }
-    Ok(None)
-}
-
-/// Per-slot initializer holing for the read-off var paths: an object init holes
-/// to its key-set form ([`hole_object_padded`], interleaved `OBJECT_PROPS`),
-/// every other init via [`hole_expr`]. The keep-shallow group path passes plain
-/// [`hole_expr`] instead (objects hole via [`hole_object`], not padded).
-fn hole_var_init_padded(init: &Expr, kept: &BTreeSet<AnchorSpan>) -> Expr {
-    match init {
-        Expr::Object(object) => Expr::Object(hole_object_padded(object, kept)),
-        other => hole_expr(other, kept),
-    }
-}
-
-/// Render a `var` binding-group selector: keep the target declarator slots (each
-/// renamed to its export via `export_for`, init holed by `hole_init` then the
-/// optional `regex_anchors` `STR_LITERAL_MATCHING_RE` post-pass), with
-/// `DECLARATORS_*` holes absorbing the runs of non-target slots. The single-target
-/// var and object read-offs are the N=1 case (one target slot, no `DECLARATORS_*`
-/// gaps).
-///
-/// The per-slot initializer holing is the one axis the call sites differ on, so
-/// it is a parameter: the read-off paths pass [`hole_var_init_padded`] (object →
-/// padded key-set holing); the keep-shallow group path passes [`hole_expr`].
-/// Factored from the near-identical `render_with` closures the object, var, and
-/// var-group minimizers each used to build inline.
-fn render_var_slots(
-    var: &VarDecl,
-    target_slots: &BTreeSet<usize>,
-    export_for: &impl Fn(&str) -> Option<String>,
-    kept: &BTreeSet<AnchorSpan>,
-    regex_anchors: &BTreeMap<AnchorSpan, String>,
-    hole_init: &impl Fn(&Expr, &BTreeSet<AnchorSpan>) -> Expr,
-) -> Result<String> {
-    let mut decls: Vec<VarDeclarator> = Vec::new();
-    let mut skipped_run = false;
-    let mut target_seen = 0usize;
-    for (idx, declarator) in var.decls.iter().enumerate() {
-        if !target_slots.contains(&idx) {
-            skipped_run = true;
-            continue;
-        }
-        if skipped_run {
-            decls.push(declarator_hole(declarator_hole_name(
-                target_seen,
-                target_slots.len(),
-            )));
-            skipped_run = false;
-        }
-        let name = single_ident_pat_name(&declarator.name)
-            .expect("target declarator is a plain identifier");
-        let mut holed = declarator.clone();
-        holed.name = named_pat(&export_for(name).expect("target declarator has an export"));
-        holed.init = declarator.init.as_ref().map(|init| {
-            let mut holed_init = hole_init(init, kept);
-            if !regex_anchors.is_empty() {
-                holed_init.visit_mut_with(&mut RegexAnchorSubstitution {
-                    patterns: regex_anchors,
-                });
-            }
-            Box::new(holed_init)
-        });
-        decls.push(holed);
-        target_seen += 1;
-    }
-    if skipped_run {
-        decls.push(declarator_hole(declarator_hole_name(
-            target_seen,
-            target_slots.len(),
-        )));
-    }
-    let mut holed_var = var.clone();
-    holed_var.declare = false;
-    holed_var.decls = decls;
-    emit_selector(ModuleItem::Stmt(Stmt::Decl(Decl::Var(Box::new(holed_var)))))
-}
-
-/// Ranked candidate anchor spans for one binding-group slot, best-first. Object
-/// slots reuse the key-set ranking ([`object_anchor_ranking`]: direct values then
-/// keys); every other init ranks its direct shallow literals first (a meaningful,
-/// rebuild-stable value pin), then structural key/member presence, then deeper
-/// literals — the keep-shallow tier order, flattened into a single best-first
-/// list the per-slot greedy consumes.
-fn slot_anchor_ranking(declarator: &VarDeclarator) -> Vec<AnchorSpan> {
-    let Some(init) = declarator.init.as_deref() else {
-        return Vec::new();
-    };
-    if let Expr::Object(object) = init {
-        return object_anchor_ranking(object);
-    }
-    let mut candidates = AnchorCandidates::default();
-    collect_expr_anchors(init, 0, &mut candidates);
-    let mut ranked: Vec<AnchorSpan> = candidates.shallow_literals();
-    for tier in candidates.deep_cover_tiers() {
-        for span in tier {
-            if !ranked.contains(&span) {
-                ranked.push(span);
-            }
-        }
-    }
-    ranked
-}
-
-/// Per-slot minimal anchor set for a binding-group read-off: the smallest kept
-/// span subset (drawn from `ranked`) that makes the matcher resolve **this slot's
-/// own export binding** to its declarator, viewed as a single-target selector
-/// (every other target slot holed to `DECLARATORS_*`). This is the per-slot
-/// declarator-tuple resolution the chunk-wide `minimal_anchor_set` (which is
-/// per-statement and resolves by distinct body index) cannot express: it cannot
-/// see two sibling declarators of the *same* statement.
-///
-/// Reuses [`cover_object_slot`]'s slot-aware scoring — `(target slot not yet
-/// resolved, total matches)` via the single-binding-form matcher
-/// ([`matched_binding_candidates`]) — so an anchor that flips the target binding
-/// to the resolved one and rules out the most competitors wins. `seed` (the
-/// chunk-wide read-off spans restricted to this slot) is kept up front when
-/// non-empty: it is the index's already-ranked `selective × stable` choice, so
-/// the greedy only adds to it. Returns the kept spans for the slot, or `None`
-/// when the slot's own anchors cannot single it out (the caller then falls back
-/// to the keep-shallow group path).
-fn slot_minimal_anchors(
-    index: &ChunkSelectorIndex,
-    var: &VarDecl,
-    decl: &IndexedDeclaration,
-    slot: usize,
-    target: &SynthesizedTargetBinding,
-    seed: &BTreeSet<AnchorSpan>,
-    ranked: &[AnchorSpan],
-) -> Result<Option<BTreeSet<AnchorSpan>>> {
-    let export = target.export_name.as_str();
-    let runtime = target.runtime_binding.as_str();
-    let slot_decl_span = var.decls[slot].span();
-    // Single-target view of this slot: the slot is the lone target, every other
-    // declarator holes to a `DECLARATORS_*` run.
-    let only_this = BTreeSet::from([slot]);
-    let export_for = |name: &str| (name == runtime).then(|| export.to_string());
-    let no_regex = BTreeMap::new();
-    let render_slot = |kept: &BTreeSet<AnchorSpan>| -> Result<String> {
-        render_var_slots(
-            var,
-            &only_this,
-            &export_for,
-            kept,
-            &no_regex,
-            &hole_var_init_padded,
-        )
-    };
-    // The slot resolves when the single-binding matcher singles out exactly this
-    // declarator: one match, at the target statement, bound to the target slot's
-    // runtime name.
-    let slot_resolves = |kept: &BTreeSet<AnchorSpan>| -> Result<bool> {
-        let matches = matched_binding_candidates(index, export, &render_slot(kept)?)?;
-        let [m] = matches.as_slice() else {
-            return Ok(false);
-        };
-        Ok(m.body_idx == decl.body_idx && m.binding.binding_name == runtime)
-    };
-    let mut kept: BTreeSet<AnchorSpan> = seed.clone();
-    while !slot_resolves(&kept)? {
-        let mut best: Option<((bool, usize), AnchorSpan)> = None;
-        for &anchor in ranked.iter().take(MAX_MINIMIZER_ANCHORS) {
-            if kept.contains(&anchor) || !node_holds_anchor(slot_decl_span, anchor) {
-                continue;
-            }
-            let mut trial = kept.clone();
-            trial.insert(anchor);
-            let matches = matched_binding_candidates(index, export, &render_slot(&trial)?)?;
-            let target_unresolved = !matches.iter().any(|m| m.binding.binding_name == runtime);
-            let score = (target_unresolved, matches.len());
-            if best.is_none_or(|(best_score, _)| score < best_score) {
-                best = Some((score, anchor));
-            }
-        }
-        let Some((_, anchor)) = best else {
-            return Ok(None);
-        };
-        kept.insert(anchor);
-    }
-    Ok(Some(kept))
-}
-
-/// Multi-target binding-group read-off (the per-slot declarator-tuple resolution
-/// the chunk-wide read-off cannot express): for each target declarator slot read
-/// its minimal anchor off the shape index (restricted to that slot) and, when the
-/// chunk-wide set does not single the slot out within the statement, extend it
-/// with a slot-aware greedy ([`slot_minimal_anchors`]). UNION the per-slot kept
-/// spans, render the whole group through [`render_var_slots`], and prove
-/// the tuple with [`prove_synthesized_selector`] (the binding-group matcher as the
-/// resolves-uniquely oracle). The regex-literal upgrade applies across all slots.
-///
-/// Returns `None` — caller falls back to the keep-shallow path — when any slot's
-/// own anchors cannot single it out or the unioned selector does not resolve the
-/// tuple uniquely.
-fn try_var_group_read_off(
-    index: &ChunkSelectorIndex,
-    var: &VarDecl,
-    decl: &IndexedDeclaration,
-    targets: &[SynthesizedTargetBinding],
-    target_slots: &BTreeSet<usize>,
-) -> Result<Option<SpecializedSelector>> {
-    let export_for = |runtime: &str| {
-        targets
-            .iter()
-            .find(|target| target.runtime_binding == runtime)
-            .map(|target| target.export_name.clone())
-    };
-    // Chunk-wide read-off anchors for the whole statement, mapped to byte spans;
-    // each slot's seed is the subset lying inside that declarator.
-    let chunk_kept: BTreeSet<AnchorSpan> = match index.shape_index.minimal_anchor_set(decl.body_idx)
-    {
-        Some(anchor_set) => {
-            let item = index
-                .parsed
-                .module
-                .body
-                .get(decl.body_idx)
-                .context("read-off body index no longer in module")?;
-            kept_spans_for_anchor_set(item, &anchor_set)
-        }
-        None => BTreeSet::new(),
-    };
-
-    let mut union: BTreeSet<AnchorSpan> = BTreeSet::new();
-    for &slot in target_slots {
-        let runtime = single_ident_pat_name(&var.decls[slot].name)
-            .expect("target declarator is a plain identifier");
-        let target = targets
-            .iter()
-            .find(|target| target.runtime_binding == runtime)
-            .expect("target slot has a target");
-        let slot_decl_span = var.decls[slot].span();
-        let seed: BTreeSet<AnchorSpan> = chunk_kept
-            .iter()
-            .copied()
-            .filter(|span| node_holds_anchor(slot_decl_span, *span))
-            .collect();
-        let ranked = slot_anchor_ranking(&var.decls[slot]);
-        let Some(slot_kept) = slot_minimal_anchors(index, var, decl, slot, target, &seed, &ranked)?
-        else {
-            return Ok(None);
-        };
-        union.extend(slot_kept);
-    }
-
-    let no_regex = BTreeMap::new();
-    // Prove the unioned tuple resolves uniquely through the binding-group matcher
-    // before offering the regex upgrade; on failure the caller falls back to the
-    // keep-shallow group path.
-    if prove_synthesized_selector(
-        index,
-        decl,
-        targets,
-        &render_var_slots(
-            var,
-            target_slots,
-            &export_for,
-            &union,
-            &no_regex,
-            &hole_var_init_padded,
-        )?,
-    )
-    .is_err()
-    {
-        return Ok(None);
-    }
-
-    // Regex-literal upgrade over the kept string literals of every target slot
-    // (shared policy with the keep-shallow path).
-    let mut regex_candidates: BTreeMap<AnchorSpan, String> = BTreeMap::new();
-    for &slot in target_slots {
-        if let Some(init) = &var.decls[slot].init {
-            regex_candidates.extend(collect_regex_anchor_candidates(init));
-        }
-    }
-    let render_with = |kept: &BTreeSet<AnchorSpan>,
-                       regex_anchors: &BTreeMap<AnchorSpan, String>|
-     -> Result<String> {
-        render_var_slots(
-            var,
-            target_slots,
-            &export_for,
-            kept,
-            regex_anchors,
-            &hole_var_init_padded,
-        )
-    };
-    let regex_anchors = accepted_regex_anchors(
-        index,
-        decl,
-        targets,
-        &regex_candidates,
-        &union,
-        &render_with,
-    )?;
-
-    let source = render_with(&union, &regex_anchors)?;
-    let rewritten_holes = holes_present(&source);
-    Ok(Some(SpecializedSelector {
-        match_source: source,
-        rewritten_holes,
-    }))
-}
-
-/// Minimal anchor cover for a `var`/`let`/`const` binding group: render the
-/// shared declaration keeping the target declarators (each `export = <holed
-/// init>`) with `DECLARATORS_*` holes for non-target runs, and pin just enough
-/// anchors for the group to resolve uniquely to the right declaration and
-/// bindings. A single-declarator target is the N=1 case of this path: one
-/// target slot, no `DECLARATORS_*` gaps, and the binding-group matcher's tuple
-/// proof degenerates to the single-binding case in `prove_synthesized_selector`.
-///
-/// The proof uses `prove_synthesized_selector` as a boolean oracle (it yields a
-/// count or an error rather than a candidate set), adding anchors tier by tier
-/// (shallow literals first) until the group resolves correctly.
-fn minimize_var_group_selector(
-    index: &ChunkSelectorIndex,
-    var: &VarDecl,
-    decl: &IndexedDeclaration,
-    targets: &[SynthesizedTargetBinding],
-) -> Result<Option<SpecializedSelector>> {
-    let export_for = |runtime: &str| {
-        targets
-            .iter()
-            .find(|target| target.runtime_binding == runtime)
-            .map(|target| target.export_name.clone())
-    };
-    let target_slots: BTreeSet<usize> = var
-        .decls
-        .iter()
-        .enumerate()
-        .filter_map(|(idx, declarator)| {
-            let name = single_ident_pat_name(&declarator.name)?;
-            export_for(name).map(|_| idx)
-        })
-        .collect();
-    if target_slots.len() != targets.len() {
-        return Ok(None);
-    }
-
-    // W3 + key-set minimization: a single-target var whose target declarator is
-    // an object literal reads its minimal selector off the shape index, then
-    // (slot-aware) covers the target object's own keys — `OBJECT_PROPS` holes
-    // around the discriminating key subset instead of keeping every key (the
-    // `object_keys_over_pinned` / key-set group over-pin). Works whether the
-    // object stands alone or sits inside a multi-declarator group. The matcher
-    // proves it (gate 1); on `None` we fall through to the keep-shallow group path
-    // below, so a target neither pass can single out is handled exactly as before.
-    if let Some(selector) = try_object_read_off(index, var, decl, targets, &target_slots)? {
-        return Ok(Some(selector));
-    }
-
-    // W2 read-off for the single-target non-object var (mirroring function/class):
-    // read the sparse `selective × stable` anchor off the shape index and hole the
-    // initializer down to it, instead of the keep-shallow path's "keep every
-    // shallow literal" over-pin. On `None` (no anchor set, anchors outside the
-    // target slot, or not uniquely resolved) we fall through to the keep-shallow
-    // group path below; multi-target groups always use that path — the read-off is
-    // single-target (per-slot tuple resolution is still the cover's job).
-    if let [target] = targets
-        && target_slots.len() == 1
-        && let Some(selector) = try_var_read_off(
-            index,
-            var,
-            decl,
-            target,
-            *target_slots.iter().next().expect("one target slot"),
-        )?
-    {
-        return Ok(Some(selector));
-    }
-
-    // Multi-target binding-group read-off: read each target declarator slot's
-    // minimal anchor off the shape index (restricted to that slot) plus a
-    // slot-aware greedy, UNION them, and prove the tuple through the binding-group
-    // matcher. This is the per-slot declarator-tuple resolution the chunk-wide
-    // read-off cannot express. On `None` (a slot the read-off cannot single out,
-    // or a non-resolving union) we fall through to the keep-shallow group path
-    // below — the same target is then handled exactly as before.
-    if let Some(selector) = try_var_group_read_off(index, var, decl, targets, &target_slots)? {
-        return Ok(Some(selector));
-    }
-
-    // Keep-shallow render via the shared var-slot renderer. Unlike the read-off
-    // paths, the keep-shallow path holes every init (objects included) through
-    // plain [`hole_expr`] — an object init holes via [`hole_object`] (list hole
-    // only where a run dropped), not the padded key-set form.
-    let render_with = |kept: &BTreeSet<AnchorSpan>,
-                       regex_anchors: &BTreeMap<AnchorSpan, String>|
-     -> Result<String> {
-        render_var_slots(
-            var,
-            &target_slots,
-            &export_for,
-            kept,
-            regex_anchors,
-            &hole_expr,
-        )
-    };
-
-    let mut candidates = AnchorCandidates::default();
-    for (idx, declarator) in var.decls.iter().enumerate() {
-        if target_slots.contains(&idx) {
-            if let Some(init) = &declarator.init {
-                collect_expr_anchors(init, 0, &mut candidates);
-            }
-        }
-    }
-
-    let no_regex = BTreeMap::new();
-    let resolves = |kept: &BTreeSet<AnchorSpan>| -> Result<bool> {
-        Ok(
-            prove_synthesized_selector(index, decl, targets, &render_with(kept, &no_regex)?)
-                .is_ok(),
-        )
-    };
-    // Always keep each slot's shallow literals (a group's declarators are its
-    // meaningful targets), then escalate to structural/deeper anchors only if
-    // the group does not yet resolve uniquely to the right bindings.
-    let mut kept: BTreeSet<AnchorSpan> = candidates.shallow_literals().into_iter().collect();
-    if !resolves(&kept)? {
-        for tier in candidates.deep_cover_tiers() {
-            kept.extend(tier);
-            if resolves(&kept)? {
-                break;
-            }
-        }
-        if !resolves(&kept)? {
-            return Ok(None);
-        }
-    }
-
-    // Regex-literal upgrade over the kept string literals of every target slot
-    // (shared with the read-off path).
-    let mut regex_candidates: BTreeMap<AnchorSpan, String> = BTreeMap::new();
-    for (idx, declarator) in var.decls.iter().enumerate() {
-        if !target_slots.contains(&idx) {
-            continue;
-        }
-        if let Some(init) = &declarator.init {
-            regex_candidates.extend(collect_regex_anchor_candidates(init));
-        }
-    }
-    let regex_anchors =
-        accepted_regex_anchors(index, decl, targets, &regex_candidates, &kept, &render_with)?;
-
-    let source = render_with(&kept, &regex_anchors)?;
-    let rewritten_holes = holes_present(&source);
-    Ok(Some(SpecializedSelector {
-        match_source: source,
-        rewritten_holes,
-    }))
-}
-
-/// Minimal anchor cover for a single-target class: render the class keeping
-/// only members (and member-body statements) that carry a chosen anchor, with
-/// `CLASS_REST` for dropped member runs, and cover sibling classes.
-fn minimize_class_selector(
-    index: &ChunkSelectorIndex,
-    class: &Class,
-    decl: &IndexedDeclaration,
-    target: &SynthesizedTargetBinding,
-) -> Result<Option<SpecializedSelector>> {
-    let export = target.export_name.clone();
-    let render_with = |kept: &BTreeSet<AnchorSpan>| -> Result<String> {
-        let mut holed_class = class.clone();
-        // A minified superclass identifier is alpha-wildcarded, so always hole
-        // `extends` to ANYTHING — it still discriminates "has a superclass" from a
-        // bare class without pinning the volatile name.
-        holed_class.super_class = class
-            .super_class
-            .as_ref()
-            .map(|_| Box::new(anything_expr()));
-        holed_class.body = hole_class_members(&class.body, kept);
-        emit_selector(ModuleItem::Stmt(Stmt::Decl(Decl::Class(ClassDecl {
-            ident: ident_node(&export),
-            declare: false,
-            class: Box::new(holed_class),
-        }))))
-    };
-    // Single-target classes read off their minimal anchor set the same way
-    // functions and objects do: the holed scaffold pins the class kind plus
-    // `extends ANYTHING`, and value anchors (a member name, a literal or callee
-    // inside a member body) map to their token spans so only the member runs
-    // carrying them survive between `CLASS_REST` holes. A class the read-off
-    // cannot single out through its own value features returns `None` and is
-    // reported as debt (never a full-AST pin).
-    render_via_read_off(index, decl, target, &render_with)
-}
-
-fn anything_param() -> Param {
-    Param {
-        span: DUMMY_SP,
-        decorators: vec![],
-        pat: anything_pat(),
-    }
-}
-
-fn named_pat(name: &str) -> Pat {
-    Pat::Ident(BindingIdent {
-        id: ident_node(name),
-        type_ann: None,
-    })
-}
-
-/// A `CLASS_REST;` class-field hole that absorbs a run of dropped members.
-fn class_rest_member() -> ClassMember {
-    ClassMember::ClassProp(ClassProp {
-        span: DUMMY_SP,
-        key: PropName::Ident(IdentName::new(CLASS_REST_HOLE_KEYWORD.into(), DUMMY_SP)),
-        value: None,
-        type_ann: None,
-        is_static: false,
-        decorators: vec![],
-        accessibility: None,
-        is_abstract: false,
-        is_optional: false,
-        is_override: false,
-        readonly: false,
-        declare: false,
-        definite: false,
-    })
-}
-
-fn hole_class_members(members: &[ClassMember], kept: &BTreeSet<AnchorSpan>) -> Vec<ClassMember> {
-    let mut out = Vec::new();
-    let mut dropped_run = false;
-    for member in members {
-        if node_retains_any(member.span(), kept) {
-            if dropped_run {
-                out.push(class_rest_member());
-                dropped_run = false;
-            }
-            out.push(hole_class_member(member, kept));
-        } else {
-            dropped_run = true;
-        }
-    }
-    if dropped_run || out.is_empty() {
-        out.push(class_rest_member());
-    }
-    out
-}
-
-fn hole_class_member(member: &ClassMember, kept: &BTreeSet<AnchorSpan>) -> ClassMember {
-    match member {
-        ClassMember::Method(m) => {
-            let mut holed = m.clone();
-            holed.function.params = m.function.params.iter().map(|_| anything_param()).collect();
-            holed.function.body = m.function.body.as_ref().map(|body| holed_block(body, kept));
-            ClassMember::Method(holed)
-        }
-        ClassMember::PrivateMethod(m) => {
-            let mut holed = m.clone();
-            holed.function.params = m.function.params.iter().map(|_| anything_param()).collect();
-            holed.function.body = m.function.body.as_ref().map(|body| holed_block(body, kept));
-            ClassMember::PrivateMethod(holed)
-        }
-        ClassMember::Constructor(ctor) => {
-            let mut holed = ctor.clone();
-            holed.params = ctor
-                .params
-                .iter()
-                .map(|_| ParamOrTsParamProp::Param(anything_param()))
-                .collect();
-            holed.body = ctor.body.as_ref().map(|body| holed_block(body, kept));
-            ClassMember::Constructor(holed)
-        }
-        // Class fields and other members carrying a kept anchor: keep verbatim.
-        _ => member.clone(),
-    }
-}
-
-/// Emit a synthesized selector item (a holed declaration) to source via the
-/// shared codegen — the only AST→string step, and the matcher's parse inverts it.
-fn emit_selector(item: ModuleItem) -> Result<String> {
-    js_ast::emit_module_source(&Module {
-        span: DUMMY_SP,
-        body: vec![item],
-        shebang: None,
-    })
-}
-
-fn finish_minimized_selector(
-    index: &ChunkSelectorIndex,
-    decl: &IndexedDeclaration,
-    target: &SynthesizedTargetBinding,
-    source: String,
-) -> Result<Option<SpecializedSelector>> {
-    let targets = std::slice::from_ref(target);
-    if prove_synthesized_selector(index, decl, targets, &source).is_err() {
-        return Ok(None);
-    }
-    let rewritten_holes = holes_present(&source);
-    Ok(Some(SpecializedSelector {
-        match_source: source,
-        rewritten_holes,
-    }))
 }
 
 fn render_var_group_selector(
@@ -4139,66 +2268,6 @@ fn is_zero(value: &usize) -> bool {
 mod selector_minimizer_proptest;
 
 #[cfg(test)]
-mod regex_anchor_pattern_tests {
-    use super::*;
-
-    #[test]
-    fn pins_stable_prefix_and_wildcards_hex_tail() {
-        let pattern = regex_anchor_pattern("chunk-a1b2c3").expect("derivable hex tail");
-        assert_eq!(pattern, "^chunk\\-[0-9A-Fa-f]+$");
-        let re = regex::Regex::new(&pattern).unwrap();
-        // The pattern matches the seen value and rebuild variants of the same
-        // stable prefix, but not a different prefix.
-        assert!(re.is_match("chunk-a1b2c3"));
-        assert!(re.is_match("chunk-ffffff"));
-        assert!(!re.is_match("widget-a1b2c3"));
-        // Anchored: a longer string sharing the prefix must not match.
-        assert!(!re.is_match("chunk-a1b2c3-extra"));
-    }
-
-    #[test]
-    fn prefers_decimal_class_for_pure_digit_tail() {
-        let pattern = regex_anchor_pattern("vendor_1024").expect("derivable digit tail");
-        assert_eq!(pattern, "^vendor_[0-9]+$");
-        let re = regex::Regex::new(&pattern).unwrap();
-        assert!(re.is_match("vendor_1024"));
-        assert!(re.is_match("vendor_9999"));
-        // The decimal class is tighter than hex: a hex-only rebuild would not
-        // match, which is fine — we wildcard only what we observed (digits).
-        assert!(!re.is_match("vendor_abcd"));
-    }
-
-    #[test]
-    fn escapes_regex_metacharacters_in_the_prefix() {
-        let pattern = regex_anchor_pattern("main.bundle.4f3a2b").expect("derivable");
-        assert_eq!(pattern, "^main\\.bundle\\.[0-9A-Fa-f]+$");
-        let re = regex::Regex::new(&pattern).unwrap();
-        assert!(re.is_match("main.bundle.4f3a2b"));
-        // The `.`s are literal, not any-char wildcards.
-        assert!(!re.is_match("mainXbundleY4f3a2b"));
-    }
-
-    #[test]
-    fn declines_short_numeric_suffixes() {
-        // A two/three-char numeric suffix is more likely meaningful than
-        // generated, so no pattern is offered (`MIN_VOLATILE_TAIL_LEN`).
-        assert_eq!(regex_anchor_pattern("v2"), None);
-        assert_eq!(regex_anchor_pattern("step3"), None);
-        assert_eq!(regex_anchor_pattern("h2o"), None);
-    }
-
-    #[test]
-    fn declines_when_no_stable_prefix_remains() {
-        // The whole literal is the volatile tail: a bare digit/hex anchor pins
-        // nothing meaningful.
-        assert_eq!(regex_anchor_pattern("123456"), None);
-        assert_eq!(regex_anchor_pattern("deadbeef"), None);
-        // Separator-only prefix is likewise rejected.
-        assert_eq!(regex_anchor_pattern("-123456"), None);
-    }
-}
-
-#[cfg(test)]
 mod adjacent_function_grouping_tests {
     use super::*;
 
@@ -4594,91 +2663,5 @@ mod prefilter_soundness_tests {
                 "the distinct-token siblings must exercise the singleton fast-path"
             );
         });
-    }
-}
-
-#[cfg(test)]
-mod interior_holing_tests {
-    use super::*;
-
-    /// Round-trip `source` through swc parse + codegen so equality is on AST
-    /// shape, not incidental formatting.
-    fn normalize(source: &str) -> String {
-        js_ast::with_swc_globals(|| {
-            let module = js_ast::parse_js_module_ast("<interior-holing>", source).unwrap();
-            js_ast::emit_module_source(&module).unwrap()
-        })
-    }
-
-    /// Span of the first string literal whose value is `needle`.
-    fn string_literal_span(expr: &Expr, needle: &str) -> Span {
-        struct Find<'a> {
-            needle: &'a str,
-            found: Option<Span>,
-        }
-        impl Visit for Find<'_> {
-            fn visit_str(&mut self, str_: &Str) {
-                if self.found.is_none() && str_.value.to_string_lossy() == self.needle {
-                    self.found = Some(str_.span);
-                }
-            }
-        }
-        let mut find = Find {
-            needle,
-            found: None,
-        };
-        expr.visit_with(&mut find);
-        find.found.expect("needle literal present in expression")
-    }
-
-    /// Hole the single expression-statement in `source`, pinning the lone
-    /// `anchor` string literal as the kept span.
-    fn hole_statement_expr(source: &str, anchor: &str) -> String {
-        js_ast::with_swc_globals(|| {
-            let module = js_ast::parse_js_module_ast("<interior-holing>", source).unwrap();
-            let [ModuleItem::Stmt(Stmt::Expr(stmt))] = module.body.as_slice() else {
-                panic!("expected a single expression statement");
-            };
-            let kept = BTreeSet::from([span_key(string_literal_span(&stmt.expr, anchor))]);
-            emit_selector(ModuleItem::Stmt(Stmt::Expr(ExprStmt {
-                span: DUMMY_SP,
-                expr: Box::new(hole_expr(&stmt.expr, &kept)),
-            })))
-            .unwrap()
-        })
-    }
-
-    #[test]
-    fn holes_nested_object_inside_a_kept_array_argument() {
-        // The object nested in a kept `ANYTHING.run([{…}])` keeps only the anchor
-        // property, holing the rest to OBJECT_PROPS. Regression guard for the
-        // `Expr::Array` recursion: the array used to fall through `hole_expr`'s
-        // verbatim catch-all, so every property was pinned.
-        let holed = hole_statement_expr(
-            r#"ctx.engine.run([{ source: ctx.node, mode: "keepMe", silent: true }]);"#,
-            "keepMe",
-        );
-        assert_eq!(
-            normalize(&holed),
-            normalize(r#"ANYTHING.run([{ OBJECT_PROPS, mode: "keepMe", OBJECT_PROPS }]);"#),
-        );
-    }
-
-    #[test]
-    fn holes_non_anchor_array_elements_to_anything() {
-        // Array elements carrying no anchor hole to ANYTHING (arity-exact, since
-        // the matcher matches array elements element-wise); only the element
-        // holding the anchor is recursed into. The lone-prop object keeps its one
-        // anchor prop with no OBJECT_PROPS padding (nothing was dropped). The bare
-        // `render` callee holes to ANYTHING (a minified name the matcher
-        // alpha-wildcards), unlike the member-method `.run` in the sibling case.
-        let holed = hole_statement_expr(
-            r#"render([first(), { mode: "keepMe" }, third()]);"#,
-            "keepMe",
-        );
-        assert_eq!(
-            normalize(&holed),
-            normalize(r#"ANYTHING([ANYTHING, { mode: "keepMe" }, ANYTHING]);"#),
-        );
     }
 }
